pyproject.toml
[tool.poetry]
name = "atomic-habits"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
black = "^25.1.0"
flake8 = "^7.1.1"
isort = "^6.0.0"
mypy = "^1.14.1"
python-dotenv = "^1.0.1"
django = "^5.1.5"
psycopg2-binary = "^2.9.10"
djangorestframework-simplejwt = "^5.4.0"
djangorestframework = "^3.15.2"
celery = "^5.4.0"
eventlet = "^0.39.0"
redis = "^5.2.1"
django-celery-beat = "^2.7.0"
requests = "^2.32.3"
django-cors-headers = "^4.6.0"
drf-yasg = "^1.21.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
line-length = 119
exclude = '.venv'

[tool.isort]
line_length = 119